Refactor the code to implement package structure instead of running the app as module
